Update README.md to refer to Python community GeoHash replication.
A little presumptuous of me to do a pull request. But this is useful for me, and I bet it will be useful for others, and I wanted to share somehow. Not sure what I should have done. Tweet at you? 

(PS I'm also an Alphabet employee, but this is my free time.)